+++ ai_detector_model/core/config.py
@@ -5,7 +5,6 @@ from loguru import logger
 
 # Paths
 PROJ_ROOT = Path(__file__).resolve().parents[1]
-logger.info(f"PROJ_ROOT path is: {PROJ_ROOT}")
 
 # Load environment variables from .env file if it exists
 load_dotenv()
@@ -36,3 +35,6 @@ try:
     logger.add(lambda msg: tqdm.write(msg, end=""), colorize=True)
 except ModuleNotFoundError:
     pass
+
+if __name__ == "__main__":
+    logger.info(f"PROJ_ROOT path is: {PROJ_ROOT}")
