+++ ai_detector_model/data/image_preprocessor.py
